arg isn't always the arguments
afaik the only way you can be sure you have a function's arguments is to
use the {...} construct. They say "arg" works, but it sort of doesn't.
In this case, calling date() with any arguments, results in an arg
variable being created (for some reason) as a table with 0 arguments.
{...} of course has the proper arguments, so I guess arg is just left in
there to break everything.

For instance, date('2014-10-24 21:56:42.095-07') gives the same result
as calling date() and never actually produces a date on the 24th of
October (unless that's the current date). It might just be luajit, but
it's probably safer to use {...} in any case.

Easy enough to explicitly specify that arg should be {...}.